+++ Packages/com.unity.render-pipelines.universal/Editor/AssetPostProcessors/MaterialPostprocessor.cs
@@ -104,7 +104,7 @@ namespace UnityEditor.Rendering.Universal
         internal static List<string> s_ImportedAssetThatNeedSaving = new List<string>();
         internal static bool s_NeedsSavingAssets = false;
 
-        internal static readonly Action<Material, ShaderID>[] k_Upgraders = { UpgradeV1, UpgradeV2, UpgradeV3, UpgradeV4, UpgradeV5, UpgradeV6 };
+        internal static readonly Action<Material, ShaderID>[] k_Upgraders = { UpgradeV1, UpgradeV2, UpgradeV3, UpgradeV4, UpgradeV5, UpgradeV6, UpgradeV7 };
 
         static internal void SaveAssetsToDisk()
         {
@@ -372,6 +372,25 @@ namespace UnityEditor.Rendering.Universal
                         BaseShaderGUI.SetMaterialKeywords(material);
                     }
                 }
+            }
+        }
+
+        // Upgrades alpha-clipped materials to include logic for automatic alpha-to-coverage support
+        static void UpgradeV7(Material material, ShaderID shaderID)
+        {
+            var surfacePropertyID = Shader.PropertyToID(Property.SurfaceType);
+            var alphaClipPropertyID = Shader.PropertyToID(Property.AlphaClip);
+            var alphaToMaskPropertyID = Shader.PropertyToID(Property.AlphaToMask);
+            if (material.HasProperty(surfacePropertyID) &&
+                material.HasProperty(alphaClipPropertyID) &&
+                material.HasProperty(alphaToMaskPropertyID))
+            {
+                bool isOpaque = material.GetFloat(surfacePropertyID) < 1.0f;
+                bool isAlphaClipEnabled = material.GetFloat(alphaClipPropertyID) > 0.0f;
+
+                float alphaToMask = (isOpaque && isAlphaClipEnabled) ? 1.0f : 0.0f;
+
+                material.SetFloat(alphaToMaskPropertyID, alphaToMask);
             }
         }
     }